Remove server side parsing of time and date.
As per the specifications for #2, "Store time/date as ISO".

Currently, the server is parsing raw dates from ISO format into more
friendly dates like "March 17th", and time with a 12 hour AM/PM format.
This commit reverts that serverside parsing, as we want to compare date
and times for purposes like removing outdated rides automatically. This
parsing will soon be done client side, so as to give back that human
readable format but allowing the server to compare times/dates.