Change ride seed to clarify purpose of entries for e2e tests
Ride seeds are become a mess when everything is merged in for many
reasons. Giving some rides last names like Past (April 10 @ 4:19),
Present (@4:20) and Future (4:21) will be useful to make tests for
time sorting, and the unique names may make it easy to find this commit
and why the rides were like this.

#4