Start edit ride functionality
The button itself cannot import information for some rides, but not
others. The problem is when a ride doesn't include unrequired fields,
and tranfering the ride from ride-list to edit-ride scrambles
the fields up for some reason.

The button to submit the editted ride to the server also does not
work.