Update ride seed by removing duplicate, reorganizing/rename fields
One entry was an exact duplicate. Removed.

Some entries still had "driver" instead of "user". Updated to 'user".

There wasn't a standard order of the last 3 fields (isDriving,
roundTrip, nonSmoking). Put every entry in that order as per the ride.ts
order to make reading easier.

One expired ride is in the seed and should be. For convenience, it has
been placed at the bottom of the file so that the ride seed can be read
to include rides that should still be on the list first.