Updated *most* e2e tests to mock g-auth and have current numbers
Each set of tests mocks g-auth sign in so that they can now be reached.
Since the ride seed keeps changing, a lot of the numbers were wrong.
They were changed for most but not all rides.

One test with many different combinations of conditions still has a
couple wrong numbers, but for the moment it's worth sharing what is
correct. The other few failing tests regard nonexistant html fields
and need a little more thought to change.