Reorder ride seed by date/time, add missing userId field to some
Some rides were missing a userId so this adds this to them. While I was
at it, I decided to make the most recent rides by date at the top. The
filtered out past ride is still the bottommost one. This also changes
userId's numbers, starting from 001 at the top and increasing 1 by 1.