Add e2e method for ride not existing with Id
elementExistsWithId() as it is set up relies on the Id to actually exist
to highlight something. For some reason, adding an else clause with no
highlighting in the event that something didn't exist would still break.
For the time being, the new method elementDoesNotExistWithId works fine.
Not as robust as one method that could do both, but there are bigger
fish to fry.