Fix validation error with edit ride
We just removed the option to change offering/requesting if you're
editing a ride. We figured that if the user is gonna change from one to
another, they might as well delete the ride and start over.
This is because of the complication of changing drive status when
you have passengers, and how it should be handled if you switch back.

closes #50