Change ngOnInit to subscribe to refreshNeeded$
To start with the refresh fix, unnessary refreshRides() are removed.
In place, on initialization there is a subscription on the new property
refreshNeeded$ in the ride list service. On subscription, refreshRides
is called. The new property is a subject<void> that is tapped when a
new ride is POSTed.

Also, adds needed semicolon to userIsAPassenger.

#47